workers/requirements.txt
cloudevents==1.10.1
functions-framework==3.5.0
google-cloud-firestore
kr8s==0.12.1
PyYAML>6
Mako==1.3.0
slack-sdk
requests
